usb.py: move shared defines here
