+++ main.py
@@ -1,3 +1,5 @@
+#!/usr/bin/env python
+
 import argparse
 import usbrply.parsers
 import usbrply.printers
