+++ usbrply/lin_pcap.py
@@ -1,9 +1,6 @@
 #!/usr/bin/env python
 
-'''
-uvusbreplay-py 01_prog2.cap
-sudo pip install pycap
-'''
+from .usb import *
 
 import pcap
 import sys
@@ -20,69 +17,6 @@ g_max_packet = float('inf')
 
 vid = 0
 pid = 0
-
-'''
-possible transfer mode
-'''
-URB_TRANSFER_IN =   0x80
-URB_ISOCHRONOUS =   0x0
-URB_INTERRUPT =     0x1
-URB_CONTROL =       0x2
-URB_BULK =          0x3
-
-'''
-possible event type
-'''
-URB_SUBMIT =        ord('S')
-URB_COMPLETE =      ord('C')
-URB_ERROR =         ord('E')
-
-urb_type2str = {
-        URB_SUBMIT: 'URB_SUBMIT',
-        URB_COMPLETE: 'URB_COMPLETE',
-        URB_ERROR: 'URB_ERROR',
-        }
-
-
-transfer2str = {
-        URB_ISOCHRONOUS: "URB_ISOCHRONOUS",
-        URB_INTERRUPT: "URB_INTERRUPT",
-        URB_CONTROL: "URB_CONTROL",
-        URB_BULK: "URB_BULK",
-        }
-
-USB_REQ_GET_STATUS =        0x00
-USB_REQ_CLEAR_FEATURE =     0x01
-# 0x02 is reserved
-USB_REQ_SET_FEATURE =       0x03
-# 0x04 is reserved
-USB_REQ_SET_ADDRESS =       0x05
-USB_REQ_GET_DESCRIPTOR =    0x06
-USB_REQ_SET_DESCRIPTOR =    0x07
-USB_REQ_GET_CONFIGURATION = 0x08
-USB_REQ_SET_CONFIGURATION = 0x09
-USB_REQ_GET_INTERFACE =     0x0A
-USB_REQ_SET_INTERFACE =     0x0B
-USB_REQ_SYNCH_FRAME =       0x0C
-
-USB_DIR_OUT =               0       # to device
-USB_DIR_IN =                0x80    # to host
-
-USB_TYPE_MASK =             (0x03 << 5)
-USB_TYPE_STANDARD =         (0x00 << 5) # 0x00
-USB_TYPE_CLASS =            (0x01 << 5) # 0x20
-USB_TYPE_VENDOR =           (0x02 << 5) # 0x40
-USB_TYPE_RESERVED =         (0x03 << 5) # 0x60
-
-USB_RECIP_MASK =            0x1f
-USB_RECIP_DEVICE =          0x00
-USB_RECIP_INTERFACE =       0x01
-USB_RECIP_ENDPOINT =        0x02
-USB_RECIP_OTHER =           0x03
-# From Wireless USB 1.0
-USB_RECIP_PORT =            0x04
-USB_RECIP_RPIPE =           0x05
-
 
 def dbg(s):
     if args and args.verbose:
@@ -174,130 +108,6 @@ def bytes2AnonArray(bytes_data, byte_type = "uint8_t"):
 def deviceStr():
     # return "dev.udev"
     return "udev"
-
-# Table 9-6. Standard Feature Selectors
-feat_i2s = {
-        # Endpoint
-        0: 'ENDPOINT_HALT',
-        # Device
-        1: 'DEVICE_REMOTE_WAKEUP',
-        # Device
-        2: 'TEST_MODE',
-        }
-
-setup_reqs = [
-            # reply type
-            #"GET_STATUS",
-            "CLEAR_FEATURE",
-            "SET_FEATURE",
-            "SET_ADDRESS",
-            "GET_DESCRIPTOR",
-            "SET_DESCRIPTOR",
-            "GET_CONFIGURATION",
-            "SET_CONFIGURATION",
-            "SET_INTERFACE",
-            "GET_INTERFACE",
-            "CLEAR_FEATURE",
-            "SYNCH_FRAME",  
-            ]
-
-def req2s(ctrl):
-    m = {
-        USB_TYPE_STANDARD: {
-            USB_REQ_GET_STATUS:         "GET_STATUS",
-            USB_REQ_CLEAR_FEATURE:      "CLEAR_FEATURE",
-            USB_REQ_SET_FEATURE:        "SET_FEATURE",
-            USB_REQ_SET_ADDRESS:        "SET_ADDRESS",
-            USB_REQ_GET_DESCRIPTOR:     "GET_DESCRIPTOR",
-            USB_REQ_SET_DESCRIPTOR:     "SET_DESCRIPTOR",
-            USB_REQ_GET_CONFIGURATION:  "GET_CONFIGURATION",
-            USB_REQ_SET_CONFIGURATION:  "SET_CONFIGURATION",
-            USB_REQ_SET_INTERFACE:      "SET_INTERFACE",
-        },
-        USB_TYPE_CLASS: {
-            USB_REQ_GET_STATUS:         "GET_STATUS",
-            USB_REQ_CLEAR_FEATURE:      "CLEAR_FEATURE",
-            USB_REQ_SET_FEATURE:        "SET_FEATURE",
-            USB_REQ_GET_INTERFACE:      "GET_INTERFACE",
-        },
-        USB_TYPE_VENDOR: {
-            USB_REQ_GET_STATUS:         "GET_STATUS",
-            USB_REQ_SET_FEATURE:        "SET_FEATURE",
-            USB_REQ_CLEAR_FEATURE:      "CLEAR_FEATURE",
-            USB_REQ_SYNCH_FRAME:        "SYNCH_FRAME",  
-        },
-    }
-    if args.fx2:
-        m[USB_TYPE_VENDOR].update({
-            0xA0:         "FX2_REG_W",
-        })
-    
-    reqType = ctrl.bRequestType & USB_TYPE_MASK
-    n = m.get(reqType, None)
-    if n is None or not ctrl.bRequest in n:
-        return None
-    reqs = n[ctrl.bRequest]
-    return reqs
-
-def req_comment(ctrl, dat):
-    # Table 9-3. Standard Device Requests
-    reqs = req2s(ctrl)
-    ret = '%s (0x%02X)' % (reqs, ctrl.bRequest)
-    if reqs == 'SET_ADDRESS':
-        ret += ': 0x%02x/%d' % (ctrl.wValue, ctrl.wValue)
-    elif reqs == 'SET_FEATURE' or reqs == 'CLEAR_FEATURE':
-        ret += ': 0x%02X (%s)' % (ctrl.wValue, feat_i2s.get(ctrl.wValue, 'unknown'))
-    elif reqs == 'FX2_REG_W':
-        addr = ctrl.wValue
-        reg2s = {
-                0xE600: 'CPUCS',
-                }
-        reg = reg2s.get(addr, None)
-        
-        ret += ': addr=0x%04X' % (ctrl.wValue)
-        if reg:
-            ret += ' (%s)' % (reg,)
-        elif addr < 0x1000:
-            ret += ' (FW load)'
-        
-        if len(dat) == 1:
-            dat = ord(dat)
-            if reg == 'CPUCS':
-                if dat & 1 == 1:
-                    ret += ', reset: hold'
-                else:
-                    ret += ', reset: release'
-    comment(ret)
-
-
-def request_type2str(bRequestType):
-    ret = "";
-    
-    if (bRequestType & USB_DIR_IN) == USB_DIR_IN:
-        ret += "USB_DIR_IN"
-    else:
-        ret += "USB_DIR_OUT"
-    
-    m = {
-        USB_TYPE_STANDARD: " | USB_TYPE_STANDARD" ,
-        USB_TYPE_CLASS: " | USB_TYPE_CLASS" ,
-        USB_TYPE_VENDOR: " | USB_TYPE_VENDOR", 
-        USB_TYPE_RESERVED: " | USB_TYPE_RESERVED",
-    }
-    ret += m[bRequestType & USB_TYPE_MASK]
-    
-    m = {
-        USB_RECIP_DEVICE: " | USB_RECIP_DEVICE",
-        USB_RECIP_INTERFACE: " | USB_RECIP_INTERFACE",
-        USB_RECIP_ENDPOINT: " | USB_RECIP_ENDPOINT",
-        USB_RECIP_OTHER: " | USB_RECIP_OTHER",
-        USB_RECIP_PORT: " | USB_RECIP_PORT",
-        USB_RECIP_RPIPE: " | USB_RECIP_RPIPE",
-    }
-    ret += m[bRequestType & USB_RECIP_MASK]
-
-    return ret;
-
 
 def print_urb(urb):
     print("URB id: 0x%016lX" % (urb.id))
@@ -490,7 +300,7 @@ class Gen:
         # Drop if is generic device management traffic
         if not args.setup and self.urb.transfer_type == URB_CONTROL:
             ctrl = usb_ctrlrequest(self.urb.ctrlrequest[0:usb_ctrlrequest_sz])
-            reqst = req2s(ctrl)
+            reqst = req2s(ctrl, fx2=args.fx2)
             if reqst in setup_reqs or reqst == "GET_STATUS" and self.urb.type == URB_SUBMIT:
                 g_pending[self.urb.id] = None
                 self.submit = None
